+++ Com.Danliris.Service.Finance.Accounting.Lib/ViewModels/CreditorAccount/CreditorAccountViewModel.cs
@@ -17,15 +17,15 @@ namespace Com.Danliris.Service.Finance.Accounting.Lib.ViewModels.CreditorAccount
 
         public string InvoiceNo { get; set; }
 
-        public long? DPP { get; set; }
+        public double? DPP { get; set; }
 
-        public long? PPN { get; set; }
+        public double? PPN { get; set; }
 
-        public long? Total { get; set; }
+        public double? Total { get; set; }
 
-        public long? Mutation { get; set; }
+        public double? Mutation { get; set; }
 
-        public long? FinalBalance { get; set; }
+        public double? FinalBalance { get; set; }
 
         public string Currency { get; set; }
